Conflicts:
	Makefile

+++ policies/DIF/RMT/Monitor/DLMonitor/DLMonitor.cc
@@ -184,7 +184,7 @@ queueStat DLMonitor::getInStat(RMTQueue * queue){
     RMTPort* port = rmtAllocator->getQueueToPortMapping(queue);
     if(port == NULL){ error("RMTPort for RMTQueue not found."); }
 
-    if(inC[port] > queue->getMaxLength()+5) {
+    if((int)inC[port] > queue->getMaxLength()+5) {
         error("no esta dropeando entrada");
     }
